refactor: split the worst mixed-responsibility files (ISA-11 item 5)
- journal.dart (5 top-level types) → journal_state.dart (SceneItem +
  JournalState), journal_mode.dart (JournalMode + notifier), and
  journal.dart (controller + provider). journal.dart re-exports the
  other two, so controllers/journal.dart stays the one import.
- share_ticket_screen.dart (640 lines, 4 responsibilities) →
  share_flow.dart (entry enum, route name, closeShareFlow),
  ticket_capture.dart (RepaintBoundary → PNG helpers, pixelRatio read
  before async gaps), share_targets.dart (Instagram/Threads/native
  integrations + Facebook App ID), widgets/share_options_sheet.dart
  (the share sheet, with the three destination tiles deduped into one
  _ShareOptionTile). The screen keeps only layout and _saveImage state.
- journaling.dart: SectionSeperator → widgets/section_separator.dart
  as SectionSeparator (spelling fixed); the duplicated discard-changes
  logic in the back button and PopScope collapsed into one
  _confirmDiscardAndPop().

Flow-symbol importers now import share_flow.dart instead of a screen.
CLAUDE.md updated. Pure moves — no behavior change.
