+++ lib/features/share/screens/share_ticket_screen.dart
@@ -1,46 +1,21 @@
 import 'dart:async';
-import 'dart:io';
-import 'dart:ui' as ui;
 
-import 'package:appinio_social_share/appinio_social_share.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gal/gal.dart';
-import 'package:share_plus/share_plus.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:movie_journal/analytics_manager.dart';
 import 'package:movie_journal/features/journal/controllers/journal.dart';
 import 'package:movie_journal/features/journal/controllers/journals.dart';
 import 'package:movie_journal/features/movie/movie_providers.dart';
+import 'package:movie_journal/features/share/share_flow.dart';
+import 'package:movie_journal/features/share/share_targets.dart';
+import 'package:movie_journal/features/share/ticket_capture.dart';
 import 'package:movie_journal/features/share/widgets/flippable_ticket.dart';
+import 'package:movie_journal/features/share/widgets/share_options_sheet.dart';
 import 'package:movie_journal/features/share/widgets/ticket_back.dart';
 import 'package:movie_journal/features/share/widgets/ticket_front.dart';
-import 'package:movie_journal/analytics_manager.dart';
 import 'package:movie_journal/features/toast/custom_toast.dart';
 import 'package:movie_journal/shared_widgets/circled_icon_button.dart';
-
-enum ShareTicketEntry { journalContent, journalComplete }
-
-/// Route name applied to every screen pushed inside the share flow
-/// (poster picker + share ticket). Used by [closeShareFlow] to pop back
-/// past all in-flow routes regardless of how many sit on the stack.
-const String kShareFlowRouteName = 'share_flow';
-
-/// Closes the share flow and returns the user to the right place based on
-/// where they entered:
-///  - [ShareTicketEntry.journalComplete] (just-saved journal): back to home.
-///  - [ShareTicketEntry.journalContent] (sharing an existing journal): back
-///    to the journal content screen below the share flow.
-void closeShareFlow(BuildContext context, ShareTicketEntry entry) {
-  final nav = Navigator.of(context);
-  switch (entry) {
-    case ShareTicketEntry.journalComplete:
-      nav.popUntil((route) => route.isFirst);
-    case ShareTicketEntry.journalContent:
-      nav.popUntil((route) => route.settings.name != kShareFlowRouteName);
-  }
-}
 
 class ShareTicketScreen extends ConsumerStatefulWidget {
   final JournalState journal;
@@ -59,8 +34,6 @@ class ShareTicketScreen extends ConsumerStatefulWidget {
 }
 
 class _ShareTicketScreenState extends ConsumerState<ShareTicketScreen> {
-  static const _facebookAppId = '1453372696513556';
-
   final _repaintKey = GlobalKey();
   bool _saving = false;
 
@@ -78,379 +51,29 @@ class _ShareTicketScreenState extends ConsumerState<ShareTicketScreen> {
   }
 
   void _showShareBottomSheet() {
-    final thoughts = widget.journal.thoughts;
-
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      backgroundColor: Theme.of(context).colorScheme.surface,
-      shape: const RoundedRectangleBorder(
-        borderRadius: BorderRadius.vertical(top: Radius.circular(20)),
-      ),
-      builder: (sheetContext) {
-        return ConstrainedBox(
-          constraints: BoxConstraints(
-            maxHeight: MediaQuery.of(sheetContext).size.height * 0.7,
-          ),
-          child: Padding(
-            padding: const EdgeInsets.fromLTRB(20, 12, 20, 40),
-            child: SingleChildScrollView(
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Center(
-                    child: Container(
-                      width: 36,
-                      height: 4,
-                      decoration: BoxDecoration(
-                        color: Colors.white.withValues(alpha: 0.3),
-                        borderRadius: BorderRadius.circular(2),
-                      ),
-                    ),
-                  ),
-                  const SizedBox(height: 20),
-                  if (thoughts.isNotEmpty) ...[
-                    const Text(
-                      'Copy text to post on Social',
-                      style: TextStyle(
-                        fontSize: 15,
-                        fontWeight: FontWeight.w600,
-                        color: Colors.white,
-                      ),
-                    ),
-                    const SizedBox(height: 12),
-                    Container(
-                      width: double.infinity,
-                      padding: const EdgeInsets.all(16),
-                      decoration: BoxDecoration(
-                        color: Colors.transparent,
-                        borderRadius: BorderRadius.circular(12),
-                        border: Border.all(
-                          color: Colors.white.withValues(alpha: 0.1),
-                          width: 1,
-                        ),
-                      ),
-                      child: Column(
-                        children: [
-                          SizedBox(
-                            width: double.infinity,
-                            child: Text(
-                              thoughts,
-                              maxLines: 10,
-                              overflow: TextOverflow.ellipsis,
-                              style: TextStyle(
-                                fontSize: 14,
-                                height: 1.5,
-                                color: Colors.white.withValues(alpha: 0.7),
-                              ),
-                            ),
-                          ),
-                          Padding(
-                            padding: const EdgeInsets.symmetric(vertical: 14),
-                            child: Divider(
-                              height: 1,
-                              color: Colors.white.withValues(alpha: 0.1),
-                            ),
-                          ),
-                          GestureDetector(
-                            // Opaque so the whole row width is tappable, not
-                            // just the centered icon+text glyphs (the default
-                            // deferToChild would ignore taps on the empty space).
-                            behavior: HitTestBehavior.opaque,
-                            onTap: () {
-                              Clipboard.setData(ClipboardData(text: thoughts));
-                              CustomToast.showSuccess(context, 'Copied to clipboard');
-                            },
-                            child: Padding(
-                              // Vertical padding enlarges the touch target so it
-                              // matches the visible button area.
-                              padding: const EdgeInsets.symmetric(vertical: 8),
-                              child: Row(
-                                mainAxisAlignment: MainAxisAlignment.center,
-                                children: [
-                                  Icon(
-                                    Icons.copy,
-                                    size: 16,
-                                    color: Colors.white.withValues(alpha: 0.7),
-                                  ),
-                                  const SizedBox(width: 6),
-                                  Text(
-                                    'Copy Text',
-                                    style: TextStyle(
-                                      fontSize: 14,
-                                      fontWeight: FontWeight.w500,
-                                      color: Colors.white.withValues(
-                                        alpha: 0.7,
-                                      ),
-                                      fontFamily: 'AvenirNext',
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                    const SizedBox(height: 24),
-                  ],
-                  const Text(
-                    'Share Option',
-                    style: TextStyle(
-                      fontSize: 15,
-                      fontWeight: FontWeight.w600,
-                      color: Colors.white,
-                    ),
-                  ),
-                  const SizedBox(height: 16),
-                  Row(
-                    children: [
-                      // Instagram Story
-                      GestureDetector(
-                        onTap: () {
-                          Navigator.of(sheetContext).pop();
-                          _shareToInstagramStory();
-                        },
-                        child: Column(
-                          mainAxisSize: MainAxisSize.min,
-                          children: [
-                            Container(
-                              width: 48,
-                              height: 48,
-                              decoration: BoxDecoration(
-                                shape: BoxShape.circle,
-                                color: Colors.transparent,
-                                border: Border.all(
-                                  color: Colors.white.withValues(alpha: 0.2),
-                                ),
-                              ),
-                              child: Center(
-                                child: Image.asset(
-                                  'assets/images/instagram_logo.png',
-                                  width: 48,
-                                  height: 48,
-                                ),
-                              ),
-                            ),
-                            const SizedBox(height: 8),
-                            const Text(
-                              'Story',
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 12,
-                                fontFamily: 'AvenirNext',
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      const SizedBox(width: 24),
-                      // Threads
-                      GestureDetector(
-                        onTap: () {
-                          Navigator.of(sheetContext).pop();
-                          _shareToThreads();
-                        },
-                        child: Column(
-                          mainAxisSize: MainAxisSize.min,
-                          children: [
-                            Container(
-                              width: 48,
-                              height: 48,
-                              decoration: BoxDecoration(
-                                shape: BoxShape.circle,
-                                color: Colors.transparent,
-                                border: Border.all(
-                                  color: Colors.white.withValues(alpha: 0.2),
-                                ),
-                              ),
-                              child: Center(
-                                child: Image.asset(
-                                  'assets/images/threads_logo.png',
-                                  width: 48,
-                                  height: 48,
-                                ),
-                              ),
-                            ),
-                            const SizedBox(height: 8),
-                            const Text(
-                              'Threads',
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 12,
-                                fontFamily: 'AvenirNext',
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      const Spacer(),
-                      // Others
-                      GestureDetector(
-                        onTap: () {
-                          Navigator.of(sheetContext).pop();
-                          _shareImageNatively();
-                        },
-                        child: Column(
-                          mainAxisSize: MainAxisSize.min,
-                          children: [
-                            Container(
-                              width: 48,
-                              height: 48,
-                              decoration: BoxDecoration(
-                                shape: BoxShape.circle,
-                                color: Colors.transparent,
-                                border: Border.all(
-                                  color: Colors.white.withValues(alpha: 0.2),
-                                ),
-                              ),
-                              child: const Icon(
-                                Icons.more_horiz,
-                                color: Colors.white,
-                                size: 24,
-                              ),
-                            ),
-                            const SizedBox(height: 8),
-                            const Text(
-                              'Others',
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 12,
-                                fontFamily: 'AvenirNext',
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  Future<Uint8List?> _captureTicketAsBytes() async {
-    final pixelRatio = MediaQuery.of(context).devicePixelRatio;
-    final boundary =
-        _repaintKey.currentContext?.findRenderObject()
-            as RenderRepaintBoundary?;
-    if (boundary == null) return null;
-    final image = await boundary.toImage(pixelRatio: pixelRatio);
-    final byteData = await image.toByteData(format: ui.ImageByteFormat.png);
-    if (byteData == null) return null;
-    return byteData.buffer.asUint8List();
-  }
-
-  Future<File?> _captureTicketToFile(String filename) async {
-    final bytes = await _captureTicketAsBytes();
-    if (bytes == null) return null;
-    final file = File('${Directory.systemTemp.path}/$filename');
-    await file.writeAsBytes(bytes);
-    return file;
-  }
-
-  Future<void> _shareToInstagramStory() async {
-    try {
-      final file = await _captureTicketToFile('movie_ticket_story.png');
-      if (file == null) return;
-
-      final socialShare = AppinioSocialShare();
-      if (Platform.isIOS) {
-        await socialShare.iOS.shareToInstagramStory(
-          _facebookAppId,
-          stickerImage: file.path,
-        );
-      } else if (Platform.isAndroid) {
-        await socialShare.android.shareToInstagramStory(
-          _facebookAppId,
-          stickerImage: file.path,
-        );
-      }
-      unawaited(
-        AnalyticsManager.logJournalShared(
-          movieTitle: widget.journal.movieTitle,
-          shareMethod: 'instagram_story',
-        ),
-      );
-    } catch (e) {
-      debugPrint('Instagram Story share error: $e');
-      if (mounted) {
-        CustomToast.showError(
-          context,
-          'Could not open Instagram. Is it installed?',
-        );
-      }
-    }
-  }
-
-  Future<void> _shareToThreads() async {
-    try {
-      final text = _composeThreadsText();
-      final uri = Uri.parse(
-        'https://www.threads.net/intent/post?text=${Uri.encodeComponent(text)}',
-      );
-
-      if (await canLaunchUrl(uri)) {
-        await launchUrl(uri, mode: LaunchMode.externalApplication);
-        unawaited(
-          AnalyticsManager.logJournalShared(
-            movieTitle: widget.journal.movieTitle,
-            shareMethod: 'threads',
-          ),
-        );
-      } else {
-        if (mounted) {
-          CustomToast.showError(
+    ShareOptionsSheet.show(
+      context,
+      thoughts: widget.journal.thoughts,
+      onInstagramStory:
+          () => shareTicketToInstagramStory(
             context,
-            'Could not open Threads. Is it installed?',
-          );
-        }
-      }
-    } catch (e) {
-      debugPrint('Threads share error: $e');
-      if (mounted) {
-        CustomToast.showError(context, 'Could not open Threads');
-      }
-    }
-  }
-
-  /// Compose the text to share on Threads.
-  /// TODO: Implement your preferred text format (~5-10 lines).
-  String _composeThreadsText() {
-    final journal = widget.journal;
-    // Available data:
-    //   journal.movieTitle  — e.g. "Fight Club"
-    //   journal.thoughts    — the user's written thoughts
-    //   journal.emotions    — List<Emotion> selected by user
-    //   journal.createdAt   — DateTime when journal was created
-    return journal.thoughts;
-  }
-
-  Future<void> _shareImageNatively() async {
-    try {
-      final file = await _captureTicketToFile('movie_ticket_share.png');
-      if (file == null) return;
-
-      await SharePlus.instance.share(ShareParams(files: [XFile(file.path)]));
-      unawaited(
-        AnalyticsManager.logJournalShared(
-          movieTitle: widget.journal.movieTitle,
-          shareMethod: 'native',
-        ),
-      );
-    } catch (e) {
-      debugPrint('Share error: $e');
-    }
+            movieTitle: widget.journal.movieTitle,
+            repaintKey: _repaintKey,
+          ),
+      onThreads: () => shareToThreads(context, journal: widget.journal),
+      onOthers:
+          () => shareTicketNatively(
+            context,
+            movieTitle: widget.journal.movieTitle,
+            repaintKey: _repaintKey,
+          ),
+    );
   }
 
   Future<void> _saveImage() async {
     if (_saving) return;
     setState(() => _saving = true);
+    final pixelRatio = MediaQuery.of(context).devicePixelRatio;
 
     try {
       // Request gallery permission
@@ -465,7 +88,10 @@ class _ShareTicketScreenState extends ConsumerState<ShareTicketScreen> {
         }
       }
 
-      final bytes = await _captureTicketAsBytes();
+      final bytes = await captureTicketAsBytes(
+        _repaintKey,
+        pixelRatio: pixelRatio,
+      );
       if (bytes == null) return;
 
       await Gal.putImageBytes(bytes);
